refactor: move gameEffects to config
ref https://github.com/mlevison/high-performance-teams-game/issues/37
