+++ src/config/gameEffects/gameEffects.spec.ts
@@ -1,0 +1,64 @@
+import { getGame } from '../../lib/testHelpers';
+
+jest.mock('../rounds', () => {
+  return {
+    rounds: {
+      1: {
+        actions: {
+          GAME_ACTION_BUILD_SERVER: {
+            type: 'ENGINEERING',
+            cost: 2,
+          },
+        },
+        effect: () => ({ capacity: 10 }),
+      },
+    },
+  };
+});
+
+describe('game effects', () => {
+  describe('Drag Effect if the team makes no engineering improvement', () => {
+    it('reduces capacity', () => {
+      const game = getGame();
+
+      // Ideally Stated as a precondition
+      expect(game.state.currentRound.number).toEqual(1);
+
+      game.nextRound();
+      expect(game.state.currentRound.number).toEqual(2);
+      expect(game.state.currentRound.capacity.total).toEqual(9);
+
+      game.nextRound();
+      expect(game.state.currentRound.number).toEqual(3);
+      expect(game.state.currentRound.capacity.total).toEqual(8);
+
+      game.nextRound();
+      expect(game.state.currentRound.number).toEqual(4);
+      expect(game.state.currentRound.capacity.total).toEqual(7);
+    });
+
+    it('stops reducing capacity when team added a BuildServer', () => {
+      const game = getGame();
+
+      // Ideally Stated as a precondition
+      expect(game.state.currentRound.number).toEqual(1);
+
+      game.nextRound();
+      expect(game.state.currentRound.number).toEqual(2);
+      expect(game.state.currentRound.capacity.total).toEqual(9);
+
+      game.selectAction('GAME_ACTION_BUILD_SERVER');
+      expect(game.state.currentRound.capacity.available).toEqual(7);
+
+      game.nextRound();
+      expect(game.state.currentRound.number).toEqual(3);
+      // Capacity unchanged because we eliminate the drag effect
+      expect(game.state.currentRound.capacity.total).toEqual(9);
+
+      game.nextRound();
+      expect(game.state.currentRound.number).toEqual(4);
+      // Capacity still **unchanged**
+      expect(game.state.currentRound.capacity.total).toEqual(9);
+    });
+  });
+});
