+++ src/config/gameEffects/index.ts
@@ -1,0 +1,1 @@
+export { gameEffects } from './gameEffects';
