+++ src/config/index.ts
@@ -3,3 +3,4 @@ export type GameActionId = GameActionIdT;
 
 export { rounds } from './rounds';
 export { gremlins } from './gremlins';
+export { gameEffects } from './gameEffects';
