+++ src/config/rounds/round2.spec.tsx
@@ -1,15 +1,15 @@
 import { getGame } from '../../lib/testHelpers';
 
 /* disable irrelevant other rounds */
-jest.mock('../index', () => ({
+jest.mock('./index', () => ({
   rounds: {
     1: require('./round1').round1,
     2: require('./round2').round2,
   },
 }));
 /* disable game effect to only tests single actions */
-jest.mock('../../state/effects/effects', () => ({
-  gameEffectList: [],
+jest.mock('../gameEffects', () => ({
+  gameEffects: [],
 }));
 
 describe('round 2', () => {
