+++ src/config/rounds/round4.spec.tsx
@@ -2,15 +2,15 @@ import { AppState } from '../../state';
 import { getGame } from '../../lib/testHelpers';
 
 /* disable irrelevant other rounds */
-jest.mock('../index', () => ({
+jest.mock('./index', () => ({
   rounds: {
     1: require('./round1').round1,
     4: require('./round4').round4,
   },
 }));
 /* disable game effect to only tests single actions */
-jest.mock('../../state/effects/effects', () => ({
-  gameEffectList: [],
+jest.mock('../gameEffects', () => ({
+  gameEffects: [],
 }));
 
 describe('round 4', () => {
