+++ src/state/effects/index.ts
@@ -1,11 +1,3 @@
-import {
-  Effect as EffectT,
-  GameEffect as GameEffectT,
-  VisibleEffect as VisibleEffectT,
-} from './types';
-export type Effect = EffectT;
-export type GameEffect = GameEffectT;
-export type VisibleEffect = VisibleEffectT;
+export type { Effect, GameEffect, VisibleEffect } from './types';
 
-export { gameEffectList } from './effects';
 export { isEffect, isVisibleEffect } from './helpers';
